Throupy (#3)
* Added basic backend functionality

-> Pulls calendar evetns from sqlite3 database
-> Displays them with a small, basic tkinter GUI
-> See the populate() function in DBHandler and add your events there.
Next step is to add an "add event" GUI interface.

* Moved folder

I'm a silly boy.

* Added Add event option and menu

-> You can now add an event with a click of a button
-> Simple GUI, needs improving
-> Next task is to remove events

* Split up pages

Created a new file for each of the application's pages.
